+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/UnusedTagsPass.php
@@ -47,6 +47,7 @@ class UnusedTagsPass implements CompilerPassInterface
         'container.service_subscriber',
         'container.stack',
         'controller.argument_value_resolver',
+        'controller.request_payload.validation_groups_resolver',
         'controller.service_arguments',
         'controller.targeted_value_resolver',
         'data_collector',